Changes as suggested by @brotkruemel
