+++ Configuration/Services.yaml
@@ -15,8 +15,8 @@ services:
    T3docs\Examples\Controller\AdminModuleController:
       tags: ['backend.controller']
 
-   T3docs\Examples\EventListener\Linkvalidator\BeforeRecordIsAnalyzedEventListener:
+   T3docs\Examples\EventListener\Linkvalidator\CheckExternalLinksToLocalPagesEventListener:
       tags:
          - name: event.listener
-           identifier: 'myExtensionBeforeRecordIsAnalyzedEventListener'
+           identifier: 'txExampleCheckExternalLinksToLocalPages'
